+++ app/_components/input/Input.module.css
@@ -6,13 +6,16 @@
     color: var(--text-g600);
     outline: none;
     border: none;
+
     &::placeholder {
         color: var(--text-disabled-300);
     }
+
     &:focus {
         color: var(--text-g800);
     }
 }
+
 .section {
     padding: 16px 24px;
     background-color: var(--background-gray-dark);
@@ -20,16 +23,4 @@
     border-radius: var(--radius-medium);
     display: flex;
     gap: 16px;
-}
-
-.label {
-    color: var(--text-g600);
-    font-size: var(--label-px);
-    line-height: var(--label-lh);
-    font-weight: var(--medium);
-}
-
-.container {
-    display: flex;
-    gap: 8px;
 }
